Refactor augment lines
