+++ src/augmenting/AugmentCategoryDisplay.tsx
@@ -1,7 +1,7 @@
 import { groupBy } from "lodash"
-import { SingleAugmentDisplay } from "./SingleAugmentDisplay"
-import { MultiAugmentDisplay } from "./MultiAugmentDisplay"
-import { Box, Stack } from "@mui/material"
+import { SingleAugmentDisplay } from "./AugmentDisplay/SingleAugmentLine"
+import { MultiAugmentDisplay } from "./AugmentDisplay/MultiAugmentLine"
+import { Stack } from "@mui/material"
 import { useAtom } from "jotai"
 import { availableAugments } from "./augmentSearchState"
 import { AugmentSearch } from "./AugmentSearch"
@@ -12,21 +12,17 @@ const CategoryPane = () => {
   const groups = groupBy(augments, (a) => (a.baseName ? a.baseName : a.name))
 
   return (
-    <Box>
-      <Box>
-        <Stack>
-          {Object.entries(groups)
-            .sort((a, b) => a[0].localeCompare(b[0]))
-            .map(([group, augments]) =>
-              augments.length === 1 ? (
-                <SingleAugmentDisplay key={group} augment={augments[0]} />
-              ) : (
-                <MultiAugmentDisplay key={group} augments={augments} />
-              ),
-            )}
-        </Stack>
-      </Box>
-    </Box>
+    <Stack>
+      {Object.entries(groups)
+        .sort((a, b) => a[0].localeCompare(b[0]))
+        .map(([group, augments]) =>
+          augments.length === 1 ? (
+            <SingleAugmentDisplay key={group} augment={augments[0]} />
+          ) : (
+            <MultiAugmentDisplay key={group} augments={augments} />
+          ),
+        )}
+    </Stack>
   )
 }
